텔레그램 수집 봇/requirements.txt
# Telegram
telethon>=1.36.0

# Database
aiosqlite>=0.20.0

# AI / ML
google-genai>=1.0.0
scikit-learn>=1.5.0
numpy>=1.26.0

# URL parsing & web scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
